Expose getPrimaryKeys() API at HollowStateEngine
